+++ apps/decodex/src/orchestrator/tests/project_layout.rs
@@ -1,7 +1,20 @@
-use crate::orchestrator::tests::{
-	self, Path, PathBuf, ReviewLevel, ServiceConfig, TEST_PROJECT_CONFIG_FILE, TempDir,
-	WorkflowDocument, fs,
+mod service_config;
+mod workflow_markdown;
+
+pub(super) use self::{
+	service_config::{
+		load_service_config, sample_service_config_toml,
+		sample_service_config_toml_with_github_command_path, service_config_dir,
+		service_config_path, service_config_toml_for_config,
+		service_config_toml_for_config_with_github_command_path,
+		service_config_with_github_token_env_var,
+		service_config_with_github_token_env_var_and_command_path,
+		service_config_with_review_level, service_workflow_path, write_service_config,
+	},
+	workflow_markdown::{profile_scoped_workflow_markdown, sample_workflow_markdown},
 };
+
+use crate::orchestrator::tests::{self, ReviewLevel, ServiceConfig, TempDir, WorkflowDocument, fs};
 
 pub(super) fn temp_project_layout() -> (TempDir, ServiceConfig, WorkflowDocument) {
 	temp_project_layout_with_tracker_project_slug_and_read_first(
@@ -13,174 +26,6 @@ pub(super) fn temp_project_layout() -> (TempDir, ServiceConfig, WorkflowDocument
 
 pub(super) fn sample_workflow() -> WorkflowDocument {
 	temp_project_layout().2
-}
-
-pub(super) fn write_service_config(repo_root: &Path, contents: &str) {
-	fs::create_dir_all(service_config_dir(repo_root)).expect("service config dir should exist");
-
-	let contents =
-		contents.replace("repo_root = \".\"", &format!("repo_root = \"{}\"", repo_root.display()));
-
-	fs::write(service_config_path(repo_root), contents).expect("service config should write");
-}
-
-pub(super) fn load_service_config(repo_root: &Path) -> ServiceConfig {
-	ServiceConfig::from_path(service_config_path(repo_root)).expect("service config should load")
-}
-
-pub(super) fn service_config_path(repo_root: &Path) -> PathBuf {
-	service_config_dir(repo_root).join(TEST_PROJECT_CONFIG_FILE)
-}
-
-pub(super) fn service_config_dir(repo_root: &Path) -> PathBuf {
-	repo_root
-		.parent()
-		.expect("repo root should have temp parent")
-		.join(".codex/decodex/projects/project")
-}
-
-pub(super) fn service_workflow_path(repo_root: &Path) -> PathBuf {
-	service_config_dir(repo_root).join("WORKFLOW.md")
-}
-
-pub(super) fn sample_service_config_toml(
-	service_id: &str,
-	tracker_api_key_env_var: &str,
-	github_token_env_var: &str,
-	worktree_root: Option<&Path>,
-	review_level: ReviewLevel,
-) -> String {
-	sample_service_config_toml_with_github_command_path(
-		service_id,
-		tracker_api_key_env_var,
-		github_token_env_var,
-		worktree_root,
-		review_level,
-		None,
-	)
-}
-
-pub(super) fn sample_service_config_toml_with_github_command_path(
-	service_id: &str,
-	tracker_api_key_env_var: &str,
-	github_token_env_var: &str,
-	worktree_root: Option<&Path>,
-	review_level: ReviewLevel,
-	github_command_path: Option<&Path>,
-) -> String {
-	let mut toml = format!(
-		r#"service_id = "{service_id}"
-
-[tracker]
-api_key_env_var = "{tracker_api_key_env_var}"
-
-[github]
-token_env_var = "{github_token_env_var}"
-"#
-	);
-
-	if let Some(github_command_path) = github_command_path {
-		toml.push_str(&format!("command_path = \"{}\"\n", github_command_path.display()));
-	}
-
-	if review_level != ReviewLevel::Strict {
-		toml.push_str("\n\n[codex]\n");
-		toml.push_str(&format!("review = \"{}\"\n", review_level.as_str()));
-	}
-
-	toml.push_str(
-		r#"
-
-[paths]
-repo_root = "."
-"#,
-	);
-
-	if let Some(worktree_root) = worktree_root {
-		toml.push_str(&format!("worktree_root = \"{}\"\n", worktree_root.display()));
-	}
-
-	toml
-}
-
-pub(super) fn service_config_toml_for_config(
-	config: &ServiceConfig,
-	github_token_env_var: &str,
-	review_level: ReviewLevel,
-) -> String {
-	service_config_toml_for_config_with_github_command_path(
-		config,
-		github_token_env_var,
-		review_level,
-		config.github().command_path(),
-	)
-}
-
-pub(super) fn service_config_toml_for_config_with_github_command_path(
-	config: &ServiceConfig,
-	github_token_env_var: &str,
-	review_level: ReviewLevel,
-	github_command_path: Option<&Path>,
-) -> String {
-	let default_worktree_root = config.repo_root().join(".worktrees");
-	let worktree_root =
-		(config.worktree_root() != default_worktree_root).then_some(config.worktree_root());
-
-	sample_service_config_toml_with_github_command_path(
-		config.service_id(),
-		config.tracker().api_key_env_var(),
-		github_token_env_var,
-		worktree_root,
-		review_level,
-		github_command_path,
-	)
-}
-
-pub(super) fn service_config_with_github_token_env_var(
-	config: &ServiceConfig,
-	token_env_var: &str,
-) -> ServiceConfig {
-	write_service_config(
-		config.repo_root(),
-		&service_config_toml_for_config(config, token_env_var, config.codex().review_level()),
-	);
-
-	load_service_config(config.repo_root())
-}
-
-pub(super) fn service_config_with_github_token_env_var_and_command_path(
-	config: &ServiceConfig,
-	token_env_var: &str,
-	github_command_path: &Path,
-) -> ServiceConfig {
-	write_service_config(
-		config.repo_root(),
-		&service_config_toml_for_config_with_github_command_path(
-			config,
-			token_env_var,
-			config.codex().review_level(),
-			Some(github_command_path),
-		),
-	);
-
-	load_service_config(config.repo_root())
-}
-
-pub(super) fn service_config_with_review_level(
-	config: &ServiceConfig,
-	review_level: ReviewLevel,
-) -> ServiceConfig {
-	write_service_config(
-		config.repo_root(),
-		&service_config_toml_for_config_with_github_command_path(
-			config,
-			config.github().token_env_var(),
-			review_level,
-			config.github().command_path(),
-		),
-	);
-
-	load_service_config(config.repo_root())
 }
 
 #[allow(dead_code)]
@@ -308,102 +153,4 @@ pub(super) fn temp_project_layout_with_workflow_markdown(
 		WorkflowDocument::from_path(config.workflow_path()).expect("workflow should load");
 
 	(temp_dir, config, workflow)
-}
-
-pub(super) fn profile_scoped_workflow_markdown(project_slug: &str) -> String {
-	let _ = project_slug;
-	let markdown = r#"
-+++
-version = 1
-
-[tracker]
-provider = "linear"
-startable_states = ["Todo"]
-terminal_states = ["Done", "Canceled", "Duplicate"]
-in_progress_state = "In Progress"
-success_state = "In Review"
-completed_state = "Done"
-failure_state = "Todo"
-opt_out_label = "decodex:manual-only"
-needs_attention_label = "decodex:needs-attention"
-
-[agent]
-transport = "stdio://"
-
-[execution]
-max_attempts = 3
-max_turns = 1
-max_retry_backoff_ms = 300000
-canonicalize_commands = ["cargo make fmt", "cargo make lint-fix"]
-verify_commands = ["cargo make check"]
-
-[execution.gate_profiles.config_subset]
-match_mode = "only"
-paths = ["config/**"]
-canonicalize_commands = []
-verify_commands = ["python3 -c 'print(\"ok\")'"]
-
-[execution.workspace_hooks]
-after_create_commands = []
-before_remove_commands = []
-timeout_seconds = 60
-
-[context]
-read_first = []
-+++
-
-Follow the repository policy.
-"#;
-
-	markdown.to_string()
-}
-
-pub(super) fn sample_workflow_markdown(
-	_project_slug: &str,
-	read_first: &[&str],
-	workflow_body: &str,
-	max_turns: u32,
-) -> String {
-	let read_first =
-		read_first.iter().map(|path| format!("\"{path}\"")).collect::<Vec<_>>().join(", ");
-	let context = format!("[context]\nread_first = [{read_first}]");
-	let markdown = format!(
-		r#"
-+++
-version = 1
-
-[tracker]
-provider = "linear"
-startable_states = ["Todo"]
-terminal_states = ["Done", "Canceled", "Duplicate"]
-in_progress_state = "In Progress"
-success_state = "In Review"
-completed_state = "Done"
-failure_state = "Todo"
-opt_out_label = "decodex:manual-only"
-needs_attention_label = "decodex:needs-attention"
-
-[agent]
-transport = "stdio://"
-
-[execution]
-max_attempts = 3
-max_turns = {max_turns}
-max_retry_backoff_ms = 300000
-gate_profiles = {{}}
-canonicalize_commands = []
-verify_commands = []
-
-[execution.workspace_hooks]
-after_create_commands = []
-before_remove_commands = []
-timeout_seconds = 60
-
-{context}
-+++
-
-{workflow_body}"#
-	);
-
-	markdown
 }
